% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghapps.R
\name{GitHub Apps}
\alias{GitHub Apps}
\alias{gh_app_token}
\alias{gh_app_installation_count}
\alias{gh_app_installation_list}
\alias{gh_app_installation_info}
\alias{gh_app_installation_delete}
\alias{gh_app_info}
\title{Authenticate as an App}
\usage{
gh_app_token(
  installation,
  app_id = Sys.getenv("GH_APP_ID"),
  app_key = Sys.getenv("GH_APP_KEY")
)

gh_app_installation_count(
  app_id = Sys.getenv("GH_APP_ID"),
  app_key = Sys.getenv("GH_APP_KEY")
)

gh_app_installation_list(
  app_id = Sys.getenv("GH_APP_ID"),
  app_key = Sys.getenv("GH_APP_KEY")
)

gh_app_installation_info(
  installation,
  app_id = Sys.getenv("GH_APP_ID"),
  app_key = Sys.getenv("GH_APP_KEY")
)

gh_app_installation_delete(
  installation,
  app_id = Sys.getenv("GH_APP_ID"),
  app_key = Sys.getenv("GH_APP_KEY")
)

gh_app_info(name = "r-universe")
}
\arguments{
\item{installation}{the target app-installation to authenticate with, This must
either be a user / organization name such as `"ropensci"`, or a repository that
has the app installed for example `"ropensci/magick"`.}

\item{app_id}{a string with the github app id}

\item{app_key}{file or string with your private key, passed to [openssl::read_key]}

\item{name}{slug of the github app from `https://github.com/apps/{name}`}
}
\value{
a temporary token that will be valid for 1 hour
}
\description{
Authenticates with the GitHub API on behalf of a GitHub app using the
app-id and private-key file that you get when registering the app. The
result is a temporary token that is valid for 1 hour, which you can use
just like other tokens to act on behalf of the app using e.g. [gh()].
}
\details{
To register a new app go to:
[https://github.com/settings/apps/new](https://github.com/settings/apps/new)
}
\examples{
\donttest{
Sys.setenv('GH_APP_ID' = 87942)
Sys.setenv('GH_APP_KEY' = "~/Downloads/r-universe.2020-11-25.private-key.pem")
gh_app_installation_count()
gh_app_installation_list()
gh_app_installation_info('jeroen')
gh_app_token('jeroen')
}
}
